Fork VM into separate repository (rewrite's history)
NB: This was done with a filter-branch command with
`--subdirectory-filter chip8_vm` followed by a rebase onto the commit where the
chip8_vm initially got it's own directory. It means that a few of the commits
that follow this one may have some confusing messages. e.g. "Move
{error,ops,vm} modules..." doesn't move anything but just updates some
visibility markers. Those commits will make more sense if you look at their
original versions in the orginal chip8-rust repository before this re-write.

The plus side to this re-write is that you can see the entire history of every
file in chip8_vm back to the initial commit. :)